refactor: 경험 카드 그리드 순환 참조 제거
